Move the "benchmark" decorator in utils
